Moved default lua program
