+++ src/ilua.c
@@ -3,6 +3,30 @@
 
 static char *moon_context_key;
 
+// TODO: put this somewhere out of the way, maybe just read
+// from `prog.lua` or something.
+const char *init_program = "\
+function init()\n \
+\tsize = 15\n \
+\tt = 0\n \
+end\n \
+\n \
+function update()\n \
+\tt = t + 0.01\n \
+\tsize = 15 * math.sin(t);\n \
+end\n \
+\n \
+function draw()\n \
+\tclear(55, 55, 225)\n \
+\tcolor(200, 150, 25)\n \
+\ttext('Hello, MoonCon User!', 1)\n \
+\trect(40, 40, size)\n \
+\n \
+\tcolor(40, 240, 120)\n \
+\trect(50 + math.cos(t*5) * 15, 50 + math.sin(t*8) * 24, 10);\n \
+end";
+
+
 int iLuaInit(MoonContext *ctx)
 {
    ctx->L = luaL_newstate();
@@ -30,7 +54,7 @@ int iLuaInit(MoonContext *ctx)
    lua_pushcfunction(ctx->L, color);
    lua_setglobal(ctx->L, "color");
 
-   if (luaL_dofile(ctx->L, "./prog.lua") == LUA_OK)
+   if (luaL_dostring(ctx->L, init_program) == LUA_OK)
    {
       printf("prog.lua: OK");
    } else {
